refactor: move images and audio to assets
